+++ src/app/(page)/new/page.tsx
@@ -1,8 +1,8 @@
 import { getServerSession } from "next-auth/next";
 
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 import Blockpage from "@/app/components/new/blockpage";
 import CreateArticleHome from "@/app/components/new/createArticleHome";
+import { authOptions } from "@/lib/auth";
 
 export default async function Page() {
   const session = await getServerSession(authOptions);
