+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,28 +1,6 @@
-import { PrismaAdapter } from "@next-auth/prisma-adapter";
-import { Session, User } from "next-auth";
 import NextAuth from "next-auth/next";
-import GithubProvider from "next-auth/providers/github";
 
-import prisma from "@/lib/prisma/client";
-
-export const authOptions = {
-  adapter: PrismaAdapter(prisma),
-  callbacks: {
-    async session({ session, user }: { session: Session; user: User }) {
-      if (session?.user) {
-        session.user.id = user.id;
-      }
-      return session;
-    },
-  },
-  providers: [
-    GithubProvider({
-      clientId: process.env.GITHUB_CLIENT_ID ?? "",
-      clientSecret: process.env.GITHUB_CLIENT_SECRET ?? "",
-    }),
-  ],
-  secret: process.env.NEXTAUTH_SECRET,
-};
+import { authOptions } from "@/lib/auth";
 
 const handler = NextAuth(authOptions);
