+++ src/lib/auth.ts
@@ -1,0 +1,24 @@
+import { PrismaAdapter } from "@next-auth/prisma-adapter";
+import { Session, User } from "next-auth";
+import GithubProvider from "next-auth/providers/github";
+
+import prisma from "@/lib/prisma/client";
+
+export const authOptions = {
+  adapter: PrismaAdapter(prisma),
+  callbacks: {
+    async session({ session, user }: { session: Session; user: User }) {
+      if (session?.user) {
+        session.user.id = user.id;
+      }
+      return session;
+    },
+  },
+  providers: [
+    GithubProvider({
+      clientId: process.env.GITHUB_CLIENT_ID ?? "",
+      clientSecret: process.env.GITHUB_CLIENT_SECRET ?? "",
+    }),
+  ],
+  secret: process.env.NEXTAUTH_SECRET,
+};
